… .vscode/settings.json
… crates/loro-internal/src/state.rs
… crates/loro-internal/src/state/tree_state.rs
… crates/loro-internal/src/txn.rs
… crates/loro-internal/src/value.rs
… crates/loro-wasm/src/lib.rs
… loro-js/src/index.ts
… loro-js/tests/event.test.ts
… loro-js/tests/misc.test.ts