This PR removed the deleted set which is used for caching all deleted TreeID. Because the complexity of querying whether a node has been deleted is only related to the node depth and the node depth is generally not too deep, we treat it as a constant term.

## Performance

|| main| PR |
|---|---|---|
| create 10^4 nodes | 4.2 ms  | 4.2 ms |
| move 10^5 times | 56.3 ms |  55.1 ms |
| checkout 10^3  | 72.9 ms | 68.6 ms |
| checkout 10^3 (300 depth) | 376.5 ms | 361.5 ms |
| realtime 10^3 move | 450 ms | 457.6 ms |

## API

When using Tree, I also found that the API design was not natural enough, so I took the advice from #223.
I added a new class `LoroTreeNode` in wasm, it takes `TreeHandler` and is responsible for operations related to the node itself.

Here are some examples
```ts
import { Loro } from 'loro-crdt'

const doc = new Loro();
// create Tree
const tree: LoroTree = doc.getTree("tree");
// create Root node
const root = tree.createNode();
// create a child node of root
const node = root.createNode();
// the unique TreeID
const id = node.id;
// make node to be another ROOT node
node.setAsRoot();
// move node to be child node of root
node.moveTo(root);
// set metadata of node
node.data.set("color", "red");
assertEquals(node.parent()!.id, root.id);
assertEquals(root.children()[0].id, node.id);
const sameNode = tree.getNodeByID(id);
if (tree.has(id)){
    // delete node
    tree.delete(id);
}
```