@zxch3n 2024-01-24
‘node.asRoot();’ sounds like it doesn't change the node, and just takes a view of it. I think setAsRoot is better