@Leeeon233 2024-01-25
> ‘node.asRoot();’ sounds like it doesn't change the node, and just takes a view of it. I think setAsRoot is better

I agree.

> > Move 10^5 times
> 
> This benchmark is moving a node to a random other node where each other node has an equal chance to be its parent, right?

Yes, 1000 nodes are created first, and randomly choose child and parent.

> Is it only local op? Why does this PR make it slower?

fixed, I will update the benchmark